+++ tango/server/workspace_server.py
@@ -49,8 +49,10 @@ class WorkspaceRequestHandler(SimpleHTTPRequestHandler):
             "step_class_name": step_info.step_class_name,
             "version": step_info.version,
             "dependencies": list(step_info.dependencies),
-            "start_time": step_info.start_time.isoformat() if step_info.start_time else None,
-            "end_time": step_info.end_time.isoformat() if step_info.end_time else None,
+            "start_time": step_info.start_time_local.isoformat()
+            if step_info.start_time_local
+            else None,
+            "end_time": step_info.end_time_local.isoformat() if step_info.end_time_local else None,
             "error": error,
             "result_location": result_location,
             "state": step_info.state.value,